update code write cloud band in files + minor fix
