+++ config/config_cbworkflow_southPacific.yml
@@ -4,8 +4,8 @@
 run_inheritance_tracking: True
 
 # Start/end date and time
-startdate: "20160115.00"
-enddate: "20160130.00"
+startdate: "20210219.00"
+enddate: "20210228.00"
 
 # Domain
 domain: "southPacific"
@@ -44,7 +44,7 @@ parameters_file: '/users/rpilon/codes/unil/cloudbandPy/parameters/parameters_sou
 saved_dirpath: '/work/FAC/FGSE/IDYST/ddomeise/default/romain/cloud_band_files' # directory where files will be saved
 
 save_dailyvar: False # save daily mean of the variable
-save_listcloudbands: False # list containing lists of cloud band (1 list per day)
+save_listcloudbands: True # list containing lists of cloud band (1 list per day)
 
 # To load files saved in saved_dirpath and not load files containing raw data
 load_saved_files: False
@@ -55,9 +55,9 @@ dir_figures: './cloud_band_figures'
 
 fig_detection_process: False
 # By default the following line allows to plot snapshots of cloud bands
-fig_time_evolution_object: True
+fig_time_evolution_object: False
 fig_time_evolution_var_cloudband: False
 fig_overlay_cloudband: False
 fig_show_bbox_around_blobs: False
-fig_inheritance_tracking: False
+fig_inheritance_tracking: True
 fig_density: False
