+++ src/cloudbandpy/cloudband.py
@@ -15,16 +15,16 @@ class CloudBand(object):
     def __init__(
         self,
         cloud_band_array: np.ndarray,
-        date_number: dt.datetime,
         area: float,
         lats: np.ndarray,
         lons: np.ndarray,
+        date_number: np.int32,
         iscloudband: bool = True,
         connected_longitudes: bool = False,
         parents: set[str] = set(),
     ):
         self.cloud_band_array = cloud_band_array
-        self.date = date_number
+        self.date_number = date_number
         self.area = area
         # return a list, here we've got only one cloud band --> [0]
         regions_props = measure.regionprops(self.cloud_band_array)[0]
@@ -35,7 +35,7 @@ class CloudBand(object):
         # Label/id of the cloud band
         self.parents = parents
         # id = "num date _ longitude (location)"
-        self.id_ = f"{date_number}_{round(self.latloncenter[1])}"
+        self.id_ = f"{self.date_number}_{round(self.latloncenter[1])}"
         #
         self.lats = lats
         self.lons = lons
@@ -56,10 +56,10 @@ class CloudBand(object):
             data = pickle.load(f)
             return cls(
                 data["cloud_band_array"],
-                data["date_number"],
                 data["area"],
                 data["lats"],
                 data["lons"],
+                data["date_number"],
                 data["iscloudband"],
                 data["connected_longitudes"],
                 data["parents"],
@@ -75,10 +75,10 @@ class CloudBand(object):
         """
         return cls(
             d["cloud_band_array"],
-            d["date_number"],
             d["area"],
             d["lats"],
             d["lons"],
+            d["date_number"],
             d["connected_longitudes"],
             d["iscloudband"],
             d["parents"],
@@ -95,10 +95,10 @@ class CloudBand(object):
             pickle.dump(
                 {
                     "cloud_band_array": self.cloud_band_array,
-                    "date_number": self.date_number,
                     "area": self.area,
                     "lats": self.lats,
                     "lons": self.lons,
+                    "date_number": self.date_number,
                     "connected_longitudes": self.connected_longitudes,
                     "iscloudband": self.iscloudband,
                     "parents": self.parents,
@@ -114,35 +114,35 @@ class CloudBand(object):
         """
         return {
             "cloud_band_array": self.cloud_band_array,
-            "date_number": self.date_number,
             "area": self.area,
             "lats": self.lats,
             "lons": self.lons,
+            "date_number": self.date_number,
             "connected_longitudes": self.connected_longitudes,
             "iscloudband": self.iscloudband,
             "parents": self.parents,
         }
 
 
-def dump_list(l, filename):
-    """
-    Dumps a list of lists of instances of `CloudBand` into a pickle file,
-    after converting the instances to dictionaries, so that `CloudBand`
-    is not pickled
+# def dump_list(l, filename):
+#     """
+#     Dumps a list of lists of instances of `CloudBand` into a pickle file,
+#     after converting the instances to dictionaries, so that `CloudBand`
+#     is not pickled
 
-    Input:
-        filename: Output file name (str)
-    """
-    with open(filename, "wb") as f:
-        pickle.dump([[c.todict() for c in j] for j in l], f)
+#     Input:
+#         filename: Output file name (str)
+#     """
+#     with open(filename, "wb") as f:
+#         pickle.dump([[c.todict() for c in j] for j in l], f)
 
 
-def load_list(filename):
-    """
-    Loads a pickle file constructed with `dump_list` into a list of
-    lists of instances of `CloudBand`
+# def load_list(filename):
+#     """
+#     Loads a pickle file constructed with `dump_list` into a list of
+#     lists of instances of `CloudBand`
 
-    Returns: list with data
-    """
-    with open(filename, "rb") as f:
-        return [[CloudBand.fromdict(e) for e in j] for j in pickle.load(f)]
+#     Returns: list with data
+#     """
+#     with open(filename, "rb") as f:
+#         return [[CloudBand.fromdict(e) for e in j] for j in pickle.load(f)]
